fix: moved xcm constants inside xcm_config
